Group structure synchronization
-Attributes needs to be set:
	groupStructureSynchronizationEnabled
	group extSource
	groups query
	group members query
	there has to be defined extSource in perun-extSources.xml

- Optional attributes:
	flatGroupStructureEnabled
	group members extSource
	group structure synchronization interval
	group structure synchronization timeout
- For this type of synchronization was created object CandidateGroup,
  which extends Group by parentGroupName (String) and extSource (ExtSource).

-Group structure synchronizations is run by scheduler every 5 minutes (by default).
 It works with threads just like member synchronization, but it has its own pool of threads and variables.
 groups are synchronized by groupName. From extSource are taken subjects with groupName, description and parentGroupName.

-Synchronization of group structure gets subject groups from extSource and creates candidatGroups from them.
 Each candidate is categorized to: group candidate to add, group to update and group to remove.

 1. Adding candidateGroups: If there is already their parentGroup (from extSource) in the structure,
	create group under that parent (CandidateGroup has parentGroupName from extSource).
	If there is no such parentGroup, create Group under baseGroup.
 2. Updating Groups: Check if, after creating new groups, there was added parent group (from extSource) of updating group.
	If there is such a group, then move updating group under that group. Then update also group description.
 3. Removing Groups: Move all subGroups under baseGroup and then remove group.

 Get list of all subGroups of basGroup and order them (leaf groups are first).
 for each subGroup in the list set neccessary attributes and run method synchronizeGroup.
 after that save information about member synchronization for that group.
 after all subGroups were proccessed, save information about group structure synchronization.

-There was also implemented forceGroupStructureSynchronization, which puts group to the first place of the pool.

-For test purposes were implemented methods getSubjectGroups() in  ExtSourceCSV, ExtSourceSql and ExtSourceSqlComplex.

-Method saveInformationAboutGroupSynchronization were duplicated. Original solution is run as new transaction,
 which was causing dataAccessException while saving information about member synchronization for each subGroup of baseGroup.
 The duplicity is run as nested transaction which works fine.

+++ perun-core/src/main/java/cz/metacentrum/perun/core/entry/GroupsManagerEntry.java
@@ -92,7 +92,7 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public void deleteGroup(PerunSession sess, Group group, boolean forceDelete) throws GroupNotExistsException, InternalErrorException, PrivilegeException, RelationExistsException, GroupAlreadyRemovedException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved {
+	public void deleteGroup(PerunSession sess, Group group, boolean forceDelete) throws GroupNotExistsException, InternalErrorException, PrivilegeException, RelationExistsException, GroupAlreadyRemovedException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved, ExternallyManagedException {
 		Utils.checkPerunSession(sess);
 		getGroupsManagerBl().checkGroupExists(sess, group);
 
@@ -102,11 +102,14 @@ public class GroupsManagerEntry implements GroupsManager {
 			throw new PrivilegeException(sess, "deleteGroup");
 				}
 
+		if(getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, group) || getGroupsManagerBl().hasGroupSynchronizedChild(sess, group))	{
+			throw new ExternallyManagedException("Group " + group + "or some of the subGroups are externally managed");
+		}
 		getGroupsManagerBl().deleteGroup(sess, group, forceDelete);
 	}
 
 	@Override
-	public void deleteGroup(PerunSession sess, Group group) throws GroupNotExistsException, InternalErrorException, PrivilegeException, RelationExistsException, GroupAlreadyRemovedException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved {
+	public void deleteGroup(PerunSession sess, Group group) throws GroupNotExistsException, InternalErrorException, PrivilegeException, RelationExistsException, GroupAlreadyRemovedException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved, ExternallyManagedException {
 		this.deleteGroup(sess, group, false);
 	}
 
@@ -125,13 +128,18 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public void deleteGroups(PerunSession perunSession, List<Group> groups, boolean forceDelete) throws GroupNotExistsException, InternalErrorException, PrivilegeException, GroupAlreadyRemovedException, RelationExistsException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved {
+	public void deleteGroups(PerunSession perunSession, List<Group> groups, boolean forceDelete) throws GroupNotExistsException, InternalErrorException, PrivilegeException, GroupAlreadyRemovedException, RelationExistsException, GroupAlreadyRemovedFromResourceException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved, ExternallyManagedException {
 		Utils.checkPerunSession(perunSession);
 		Utils.notNull(groups, "groups");
 
 		//Test if all groups exists and user has right to delete all of them
 		for(Group group: groups) {
 			getGroupsManagerBl().checkGroupExists(perunSession, group);
+
+			if(getGroupsManagerBl().isGroupInStructureSynchronizationTree(perunSession, group) || getGroupsManagerBl().hasGroupSynchronizedChild(perunSession, group))	{
+				throw new ExternallyManagedException("Group " + group + "or som of the subGroups are externally managed!");
+			}
+
 			//test of privileges on group
 			if(!AuthzResolver.isAuthorized(perunSession, Role.VOADMIN, group) && !AuthzResolver.isAuthorized(perunSession, Role.GROUPADMIN, group)) {
 				throw new PrivilegeException(perunSession, "deleteGroups");
@@ -162,13 +170,22 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public void moveGroup(PerunSession sess, Group destinationGroup, Group movingGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupMoveNotAllowedException, WrongAttributeValueException, WrongReferenceAttributeValueException{
+	public void moveGroup(PerunSession sess, Group destinationGroup, Group movingGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupMoveNotAllowedException, WrongAttributeValueException, WrongReferenceAttributeValueException, ExternallyManagedException, AttributeNotExistsException, WrongAttributeAssignmentException {
 		Utils.checkPerunSession(sess);
 
 		getGroupsManagerBl().checkGroupExists(sess, movingGroup);
+
+		if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, movingGroup)) {
+			throw new ExternallyManagedException("Moving group: " + movingGroup + " is externally managed!");
+		}
 		//if destination group is null, moving group will be moved as top level group
 		if(destinationGroup != null){
 			getGroupsManagerBl().checkGroupExists(sess, destinationGroup);
+
+			Attribute attrSynchronizeEnabled = getPerunBl().getAttributesManagerBl().getAttribute(sess, destinationGroup, GROUPSSTRUCTURESYNCHROENABLED_ATTRNAME);
+			if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, destinationGroup) || Objects.equals("true", (String) attrSynchronizeEnabled.getValue())) {
+				throw new ExternallyManagedException("Destination group: " + destinationGroup + " is externally managed!");
+			}
 
 			// Authorization (destination group is not null)
 			if ((!AuthzResolver.isAuthorized(sess, Role.VOADMIN, movingGroup) && !AuthzResolver.isAuthorized(sess, Role.GROUPADMIN, movingGroup)) ||
@@ -244,7 +261,8 @@ public class GroupsManagerEntry implements GroupsManager {
 
 		// Check if the group is externally synchronized
 		Attribute attrSynchronizeEnabled = getPerunBl().getAttributesManagerBl().getAttribute(sess, group, GROUPSYNCHROENABLED_ATTRNAME);
-		if (Objects.equals("true", (String) attrSynchronizeEnabled.getValue())) {
+		Attribute structureSynchronizeEnabled = getPerunBl().getAttributesManagerBl().getAttribute(sess, group, GROUPSSTRUCTURESYNCHROENABLED_ATTRNAME);
+		if (Objects.equals("true", (String) attrSynchronizeEnabled.getValue()) || getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, group) || Objects.equals(true, structureSynchronizeEnabled.getValue())) {
 			throw new ExternallyManagedException("Adding of member is not allowed. Group is externally managed.");
 		}
 
@@ -265,7 +283,8 @@ public class GroupsManagerEntry implements GroupsManager {
 
 		// Check if the group is externally synchronized
 		Attribute attrSynchronizeEnabled = getPerunBl().getAttributesManagerBl().getAttribute(sess, group, GROUPSYNCHROENABLED_ATTRNAME);
-		if (Objects.equals("true", (String) attrSynchronizeEnabled.getValue())) {
+		Attribute structureSynchronizeEnabled = getPerunBl().getAttributesManagerBl().getAttribute(sess, group, GROUPSSTRUCTURESYNCHROENABLED_ATTRNAME);
+		if (Objects.equals("true", (String) attrSynchronizeEnabled.getValue()) || getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, group) || Objects.equals(true, structureSynchronizeEnabled.getValue())) {
 			throw new ExternallyManagedException("Removing of member is not allowed. Group is externally managed.");
 		}
 
@@ -961,6 +980,20 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
+    public void forceGroupStructureSynchronization(PerunSession sess, Group group) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupStructureSynchronizationAlreadyRunningException {
+        Utils.checkPerunSession(sess);
+        getGroupsManagerBl().checkGroupExists(sess, group);
+
+        // Authorization
+        if (!AuthzResolver.isAuthorized(sess, Role.VOADMIN, group)
+                && !AuthzResolver.isAuthorized(sess, Role.GROUPADMIN, group))  {
+            throw new PrivilegeException(sess, "synchronizeGroupStructure");
+        }
+
+        getGroupsManagerBl().forceGroupStructureSynchronization(sess, group);
+    }
+
+	@Override
 	public void synchronizeGroups(PerunSession sess) throws InternalErrorException, PrivilegeException {
 		Utils.checkPerunSession(sess);
 
@@ -973,7 +1006,19 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public List<Group> getMemberGroups(PerunSession sess, Member member) throws InternalErrorException, PrivilegeException, MemberNotExistsException {
+	public void synchronizeGroupsStructures(PerunSession sess) throws InternalErrorException, PrivilegeException {
+		Utils.checkPerunSession(sess);
+
+		// Authorization
+		if (!AuthzResolver.isAuthorized(sess, Role.PERUNADMIN))  {
+			throw new PrivilegeException(sess, "synchronizeGroupsStructures");
+		}
+
+		getGroupsManagerBl().synchronizeGroupsStructures(sess);
+	}
+
+		@Override
+		public List<Group> getMemberGroups(PerunSession sess, Member member) throws InternalErrorException, PrivilegeException, MemberNotExistsException {
 		Utils.checkPerunSession(sess);
 		getPerunBl().getMembersManagerBl().checkMemberExists(sess, member);
 
@@ -1243,10 +1288,16 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public Group createGroupUnion(PerunSession sess, Group resultGroup, Group operandGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupRelationNotAllowed, GroupRelationAlreadyExists, WrongAttributeValueException, WrongReferenceAttributeValueException {
+	public Group createGroupUnion(PerunSession sess, Group resultGroup, Group operandGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupRelationNotAllowed, GroupRelationAlreadyExists, WrongAttributeValueException, WrongReferenceAttributeValueException, ExternallyManagedException {
 		Utils.checkPerunSession(sess);
 		getGroupsManagerBl().checkGroupExists(sess, resultGroup);
 		getGroupsManagerBl().checkGroupExists(sess, operandGroup);
+
+		if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, resultGroup)) {
+			throw new ExternallyManagedException("Result group: " + resultGroup + " is externally managed!");
+		} else if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, operandGroup)) {
+			throw new ExternallyManagedException("Operand group: " + operandGroup + " is externally managed!");
+		}
 
 		// Authorization
 		if ((!AuthzResolver.isAuthorized(sess, Role.VOADMIN, resultGroup) && !AuthzResolver.isAuthorized(sess, Role.GROUPADMIN, resultGroup)) ||
@@ -1258,10 +1309,16 @@ public class GroupsManagerEntry implements GroupsManager {
 	}
 
 	@Override
-	public void removeGroupUnion(PerunSession sess, Group resultGroup, Group operandGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved {
+	public void removeGroupUnion(PerunSession sess, Group resultGroup, Group operandGroup) throws InternalErrorException, GroupNotExistsException, PrivilegeException, GroupRelationDoesNotExist, GroupRelationCannotBeRemoved, ExternallyManagedException {
 		Utils.checkPerunSession(sess);
 		getGroupsManagerBl().checkGroupExists(sess, resultGroup);
 		getGroupsManagerBl().checkGroupExists(sess, operandGroup);
+
+		if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, resultGroup)) {
+			throw new ExternallyManagedException("Result group: " + resultGroup + " is externally managed!");
+		} else if (getGroupsManagerBl().isGroupInStructureSynchronizationTree(sess, operandGroup)) {
+			throw new ExternallyManagedException("Operand group: " + operandGroup + " is externally managed!");
+		}
 
 		// Authorization
 		if ( (!AuthzResolver.isAuthorized(sess, Role.VOADMIN, resultGroup) && !AuthzResolver.isAuthorized(sess, Role.GROUPADMIN, resultGroup)) ||
